app/schedule_service/serializers.py
from rest_framework import serializers

from github_integration.models import Repository


class RepositorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Repository
        fields = '__all__'
